requirements.txt
# core data tools
pandas>=1.5
requests>=2.28

# social sentiment script
tweepy>=4.0            # Twitter v2 client
snscrape>=0.4          # tweet scraping fallback
pytrends>=4.8          # Google Trends

# Solana on-chain for macro script
solana>=0.27           # solana-py RPC client

# (optional) env-var loader
python-dotenv>=1.0
